Tighten project typing and extract project internals
Refactor the project subsystem so OrisonProject coordinates smaller helper modules for module loading and render orchestration.

Add Zod validation at runtime boundaries for JSON data, list page outputs, and callable module exports while removing explicit any usage from the project slice.

Increase lint and TypeScript strictness, bump the package to 2.3.0, and verify the build with npm run test and npm run fix.
